# Speed up suggestion insertion by using a single transaction